fix(viz): north-up raster, CRS axis order, DTW/band/correlation correctness
- native_factor_map: flip grid rows once at transfer (north-up) so painter
  canvases, the QGIS GeoTIFF mirror, and raster exports agree with world
  coordinates instead of rendering factor maps N-S mirrored (F1/P1).
- mapping/reference_layers: pin traditional GIS axis order on both transform
  sides so EPSG:4326 (default) reference layers stop mirroring across the
  diagonal (F2/P1).
- viz/dtw_log_matcher: impute LAS nulls before z-normalize (NaN poisoned the
  cost matrix) and bound the O(n*m) cost matrix via decimation for long curves
  (E3/E8).
- viz/welllog_engine_adapter: diff markers in update_plan_to_view (patch
  cannot carry markers; force full resubmit) and vendor FACIES_COLORS so the
  geoviz-facade-only import policy holds and lithology fills apply (E1 + 3
  baseline test fixes).
- viz/welllog_multi_well_adapter: order correlation bands by depth, not label
  sort order, eliminating inverted 'bowtie' bands (E2).
- viz/seismic_volume_source: detect degenerate/unstructured SEGY geometry,
  guard slice reads, and invalidate the shared source on file change (D2/D3).
- viz/joint_host: defer L1 LOD start to the worker's terminal release, hoist
  TD-table parsing out of the per-line loop, add release/cancel guards (D5/D6).
- viz/seismic_3d_api: import Any, rename the AttributeTaskWorker signal off
  QThread.finished (D7/D8).
- viz/formation_volume: accumulate in float64 to avoid UTM-offset cancellation
  (E10); viz/formation_top_correlator: use the curve's real depth axis for
  DTW transfer (E9); viz/geomodel/analysis: guard empty-layer auto-tie (D14).
- viz/well_section_datum: emit diagnostics for un-correctable wells (E11);
  hosts: reproducible well-tie seed, apply datum mode after combo rebuild.
- mapping/map_interaction: even-odd hole semantics for point-in-polygon;
  map_render_backend: queue pre-bridge snapshot so the first async frame is
  not blank; interpretation_lifecycle: drop dead CRS computations.
- viz/ipc SharedMemoryArrayHandle: release mappings and clear the array ref.

+++ paleo_workbench/viz/formation_volume.py
@@ -55,7 +55,9 @@ class FormationVolumeIntegrator:
         """Construct watertight closed 3D polyhedron and compute surface integral via Gauss Divergence Theorem."""
         # Build 3D mesh vertices: top vertices [0..N-1], bot vertices [N..2N-1]
         n = rows * cols
-        v_total = np.vstack([top_verts, bot_verts])
+        # Accumulate in float64: float32 cross products cancel catastrophically
+        # at UTM-scale coordinates (verified >2000% volume error).
+        v_total = np.vstack([top_verts, bot_verts]).astype(np.float64)
 
         faces = []
